fix signedness for geline return value

+++ src/server/common.cc
@@ -64,8 +64,8 @@ WebGUI::execute(std::string &cmd, std::string &stdout, bool read)
       fflush(f);
       char *buf = NULL;
       size_t len = 0;
-      size_t read_len = 0;
-      while ((read_len = getline(&buf, &len, f)) != (unsigned)-1) {
+      ssize_t read_len = 0;
+      while ((read_len = getline(&buf, &len, f)) != -1) {
 	//	cout << "WebGUI::execute(): " << string(buf) << ", " << len << ", " << read_len << endl;
 	stdout += string(buf) + " ";
       }
